LPD-52638 Moving object-test for Jakarta upgrade to the proper class
