--- modules/apps/object/object-test/src/testIntegration/java/com/liferay/object/internal/upgrade/v11_0_0/test/ObjectJakartaUpgradeProcessTest.java
@@ -1,212 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.object.internal.upgrade.v11_0_0.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.object.constants.ObjectActionExecutorConstants;
-import com.liferay.object.constants.ObjectActionTriggerConstants;
-import com.liferay.object.constants.ObjectValidationRuleConstants;
-import com.liferay.object.field.builder.TextObjectFieldBuilder;
-import com.liferay.object.model.ObjectAction;
-import com.liferay.object.model.ObjectDefinition;
-import com.liferay.object.model.ObjectValidationRule;
-import com.liferay.object.service.ObjectActionLocalService;
-import com.liferay.object.service.ObjectDefinitionLocalService;
-import com.liferay.object.service.ObjectValidationRuleService;
-import com.liferay.object.test.util.ObjectDefinitionTestUtil;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.cache.MultiVMPool;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.test.TestInfo;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.upgrade.UpgradeStep;
-import com.liferay.portal.kernel.util.UnicodePropertiesBuilder;
-import com.liferay.portal.security.script.management.test.util.ScriptManagementConfigurationTestUtil;
-import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.portal.vulcan.util.LocalizedMapUtil;
-
-import java.util.Collections;
-import java.util.Objects;
-
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * @author Luis Ortiz
- */
-@RunWith(Arquillian.class)
-public class ObjectJakartaUpgradeProcessTest {
-
-	@ClassRule
-	@Rule
-	public static final LiferayIntegrationTestRule liferayIntegrationTestRule =
-		new LiferayIntegrationTestRule();
-
-	@Before
-	public void setUp() throws Exception {
-		_originalName = PrincipalThreadLocal.getName();
-		_originalPermissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		_user = TestPropsValues.getUser();
-
-		PermissionThreadLocal.setPermissionChecker(
-			PermissionCheckerFactoryUtil.create(_user));
-
-		PrincipalThreadLocal.setName(_user.getUserId());
-
-		_upgradeStepRegistrator.register(
-			(fromSchemaVersionString, toSchemaVersionString, upgradeSteps) -> {
-				for (UpgradeStep upgradeStep : upgradeSteps) {
-					Class<?> clazz = upgradeStep.getClass();
-
-					if (Objects.equals(
-							clazz.getName(),
-							"com.liferay.object.internal.upgrade.v11_0_0." +
-								"ObjectJakartaUpgradeProcess")) {
-
-						_upgradeProcess = (UpgradeProcess)upgradeStep;
-					}
-				}
-			});
-
-		ScriptManagementConfigurationTestUtil.save(true);
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		PermissionThreadLocal.setPermissionChecker(_originalPermissionChecker);
-
-		PrincipalThreadLocal.setName(_originalName);
-
-		ScriptManagementConfigurationTestUtil.save(false);
-	}
-
-	@Test
-	@TestInfo("LPD-52638")
-	public void testUpgrade() throws Exception {
-		ObjectAction objectAction = null;
-		ObjectDefinition objectDefinition = null;
-		ObjectValidationRule objectValidationRule = null;
-
-		try {
-			objectDefinition = ObjectDefinitionTestUtil.publishObjectDefinition(
-				Collections.singletonList(
-					new TextObjectFieldBuilder(
-					).labelMap(
-						LocalizedMapUtil.getLocalizedMap(
-							RandomTestUtil.randomString())
-					).name(
-						"a" + RandomTestUtil.randomString()
-					).build()));
-
-			objectAction = _objectActionLocalService.addObjectAction(
-				RandomTestUtil.randomString(), TestPropsValues.getUserId(),
-				objectDefinition.getObjectDefinitionId(), true,
-				StringPool.BLANK, RandomTestUtil.randomString(),
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				LocalizedMapUtil.getLocalizedMap(RandomTestUtil.randomString()),
-				RandomTestUtil.randomString(),
-				ObjectActionExecutorConstants.KEY_GROOVY,
-				ObjectActionTriggerConstants.KEY_STANDALONE,
-				UnicodePropertiesBuilder.put(
-					"script", _JAVAX_SCRIPT
-				).build(),
-				false);
-
-			objectValidationRule =
-				_objectValidationRuleService.addObjectValidationRule(
-					StringPool.BLANK, objectDefinition.getObjectDefinitionId(),
-					true, ObjectValidationRuleConstants.ENGINE_TYPE_GROOVY,
-					LocalizedMapUtil.getLocalizedMap(
-						RandomTestUtil.randomString()),
-					LocalizedMapUtil.getLocalizedMap(
-						RandomTestUtil.randomString()),
-					ObjectValidationRuleConstants.OUTPUT_TYPE_FULL_VALIDATION,
-					_JAVAX_SCRIPT, false, Collections.emptyList());
-
-			_upgradeProcess.upgrade();
-
-			_multiVMPool.clear();
-
-			ObjectAction updatedObjectAction =
-				_objectActionLocalService.getObjectAction(
-					objectAction.getObjectActionId());
-
-			Assert.assertNotNull(updatedObjectAction);
-
-			Assert.assertEquals(
-				"script=" + _JAKARTA_SCRIPT + "\n",
-				updatedObjectAction.getParameters());
-
-			ObjectValidationRule updatedObjectValidationRule =
-				_objectValidationRuleService.getObjectValidationRule(
-					objectValidationRule.getObjectValidationRuleId());
-
-			Assert.assertNotNull(updatedObjectValidationRule);
-
-			Assert.assertEquals(
-				_JAKARTA_SCRIPT, updatedObjectValidationRule.getScript());
-		}
-		finally {
-			if (objectAction != null) {
-				_objectActionLocalService.deleteObjectAction(objectAction);
-			}
-
-			if (objectValidationRule != null) {
-				_objectValidationRuleService.deleteObjectValidationRule(
-					objectValidationRule.getObjectValidationRuleId());
-			}
-
-			if (objectDefinition != null) {
-				_objectDefinitionLocalService.deleteObjectDefinition(
-					objectDefinition);
-			}
-		}
-	}
-
-	private static final String _JAKARTA_SCRIPT =
-		"System.out.println(\"import jakarta.servlet.GenericServlet\");";
-
-	private static final String _JAVAX_SCRIPT =
-		"System.out.println(\"import javax.servlet.GenericServlet\");";
-
-	@Inject(
-		filter = "component.name=com.liferay.object.internal.upgrade.registry.ObjectServiceUpgradeStepRegistrator"
-	)
-	private static UpgradeStepRegistrator _upgradeStepRegistrator;
-
-	@Inject
-	private MultiVMPool _multiVMPool;
-
-	@Inject
-	private ObjectActionLocalService _objectActionLocalService;
-
-	@Inject
-	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-	@Inject
-	private ObjectValidationRuleService _objectValidationRuleService;
-
-	private String _originalName;
-	private PermissionChecker _originalPermissionChecker;
-	private UpgradeProcess _upgradeProcess;
-	private User _user;
-
-}
